Red after major refactoring on board header and souce

+++ include/board.hpp
@@ -18,7 +18,7 @@ class Board
 {
   private:
   static shared_ptr<Board> _board_table; //Points to itself after initialized
-  static const unsigned int _size_of_table = 8; //Number of rows and ranks
+  static const uint8_t _size_of_table = 8; //Number of rows and ranks
   PBoardSquare _board_square_matrix[_size_of_table][_size_of_table]; //Matrix to store the board squares
 
   public:
@@ -30,16 +30,16 @@ class Board
 
   /*! Returns the board square on given coordinates
   */
-  PBoardSquare getBoardSquareAt(unsigned int rank_coordinate, unsigned int file_coordinate );
+  PBoardSquare getBoardSquareAt(uint8_t rank_coordinate, uint8_t file_coordinate );
 
-  unsigned int setPieceAt( unsigned int rank_coordinate, 
-                unsigned int file_coordinate , PPiece piece_to_be_set );
+  uint8_t setPieceAt( uint8_t rank_coordinate, 
+                uint8_t file_coordinate , PPiece piece_to_be_set );
 
-  bool isClearHorizontal(unsigned int actual_rank_coordinate, 
-              unsigned int actual_file_coordinate, unsigned int future_file_coordinate) const;
-              
-  bool isClearVertical(unsigned int actual_rank_coordinate,
-               unsigned int actual_file_coordinate, unsigned int future_rank_coordinate) const;
+  bool isClearHorizontal(uint8_t actual_rank_coordinate, 
+              uint8_t actual_file_coordinate, uint8_t future_file_coordinate) const;
+
+  bool isClearVertical(uint8_t actual_rank_coordinate,
+               uint8_t actual_file_coordinate, uint8_t future_rank_coordinate) const;
 };
 
 #endif
